pyproject.toml
[project]
name = "openrouter_experimentation_project"
version = "0.0.0"
description = "To experiment with OpenRouter."
readme = "coming"
requires-python = ">=3.12,<3.13"
dependencies = [
    "python-dotenv~=1.2.0",
    "requests",
]

[dependency-groups]
local = []
staging = []
prod = []
